Fixed bugs in NFT and Avatar API, Added  CollectNFT, CollectGeoNFT methods to NFTManager and NFTController,  changed Lat and Long to double for GeoNFTs,

+++ OASIS Architecture/NextGenSoftware.OASIS.API.Core/Interfaces/NFT/GeoSpatialNFT/IWeb4GeoSpatialNFT.cs
@@ -11,8 +11,8 @@ namespace NextGenSoftware.OASIS.API.Core.Interfaces.NFT.GeoSpatialNFT
         EnumValue<ProviderType> GeoNFTMetaDataProvider { get; set; }
         //EnumValue<ProviderType> OriginalOASISNFTProviderType { get; set; }
         DateTime PlacedOn { get; set; }
-        long Lat { get; set; }
-        long Long { get; set; }
+        double Lat { get; set; }
+        double Long { get; set; }
         bool AllowOtherPlayersToAlsoCollect { get; set; }
         bool PermSpawn { get; set; }
         int GlobalSpawnQuantity { get; set; }
